Rename MemorySourceError into ContentError and make it official.
