+++ src/decode/error.rs
@@ -30,8 +30,64 @@ pub trait Error: Sized + error::Error {
     fn unimplemented<T: fmt::Display + Send + Sync + 'static>(
         msg: T
     ) -> Self;
-
-    /// Converts the error into a different decoding error.
-    fn convert_into<E: Error>(self) -> E;
 }
+
+
+//------------ ContentError --------------------------------------------------
+
+/// An error for a source where the content can be wrong.
+///
+/// This is an `Error` type for all sources where acquiring the raw data
+/// cannot fail. All errors need to be able to convert from this type.
+pub struct ContentError {
+    kind: ErrorKind,
+    msg: Box<dyn fmt::Display + Send + Sync>,
+}
+
+#[derive(Clone, Copy, Debug)]
+enum ErrorKind {
+    Malformed,
+    Unimplemented,
+}
+
+impl Error for ContentError {
+    fn malformed<T: fmt::Display + Send + Sync + 'static>(
+        msg: T
+    ) -> Self {
+        ContentError {
+            kind: ErrorKind::Malformed,
+            msg: Box::new(msg)
+        }
+    }
+
+    fn unimplemented<T: fmt::Display + Send + Sync + 'static>(
+        msg: T
+    ) -> Self {
+        ContentError {
+            kind: ErrorKind::Unimplemented,
+            msg: Box::new(msg)
+        }
+    }
+}
+
+impl fmt::Debug for ContentError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("ContentError")
+            .field("kind", &self.kind)
+            .field("msg", &format_args!("{}", &self.msg))
+            .finish()
+    }
+}
+
+impl fmt::Display for ContentError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self.kind {
+            ErrorKind::Malformed => write!(f, "malformed data")?,
+            ErrorKind::Unimplemented => write!(f, "format not implemented")?,
+        }
+        write!(f, ": {}", self.msg)
+    }
+}
+
+impl error::Error for ContentError { }
